backend/pyproject.toml
[project]
name = "python-sdk-example-backend"
version = "0.1.0"
description = "Backend API for Wikipedia article generation"
authors = [
    { name = "Latitude", email = "manu@latitude.so" }
]
requires-python = ">=3.11,<3.15"
dependencies = [
    "latitude-sdk==5.6.0",
    "promptl-ai==1.3.0",
    "latitude-telemetry==2.0.2",
    "openai==2.16.0",
    "python-dotenv==1.2.1",
    "fastapi==0.128.0",
    "uvicorn==0.39.0",
]

[tool.pyright]
venvPath = "."
venv = ".venv"
